Refactor Modal - remove default Confirm/Cancel, fix default props, fix docs
