+++ packages/orcs-design-system/lib/components/Modal/Modal.mdx
@@ -1,0 +1,28 @@
+import Modal from ".";
+import { Story, Preview, Props } from "@storybook/addon-docs/blocks";
+
+# Modal
+
+[Properties](#properties)
+
+The `Modal` component brings up a popup container upon clicking on a component that changes the `visible` property of the `Modal` to `true`, for example a `Button` or even a `StyledLink`.
+
+This is a blank slate and functions to close the `Modal` and take any action must be specified (including any `Button` elements contained in the popup). The only element included by default is the close icon `X` on the upper right corner, whose `onClose` function **must** be specified. See Properties table below.
+
+For a special case of this that has Confirm and Cancel functionality baked in, see [Dialogue](./?path=/docs/components-dialogue--basic).
+
+## Default
+
+<Preview>
+  <Story id="components-modal--basic" />
+</Preview>
+
+## Using Link
+
+<Preview>
+  <Story id="components-modal--link" />
+</Preview>
+
+## Properties
+
+<Props of={Modal} />
