+++ packages/orcs-design-system/lib/components/Modal/Modal.stories.js
@@ -1,18 +1,29 @@
 import React, { useState } from "react";
 import Button from "../Button";
 import Modal from ".";
+import StyledLink from "../StyledLink";
+import Divider from "../Divider";
+import { H2, P } from "../Typography";
+import Box from "../Box";
+import Spacer from "../Spacer";
+import mdx from "./Modal.mdx";
 
 export default {
-  title: "Units/Modal",
+  title: "Components/Modal",
+  parameters: {
+    docs: {
+      page: mdx
+    }
+  },
   component: Modal
 };
 
-export const BasicModal = () => {
+export const Basic = () => {
   const [visible, setVisible] = useState(false);
   const handleOnButtonClick = () => {
     setVisible(true);
   };
-  const handleOnCancel = () => {
+  const handleOnClose = () => {
     setVisible(false);
   };
   const handleOnConfirm = () => {
@@ -24,7 +35,7 @@ export const BasicModal = () => {
       <Modal
         visible={visible}
         onConfirm={handleOnConfirm}
-        onCancel={handleOnCancel}
+        onClose={handleOnClose}
       >
         Content of the modal
       </Modal>
@@ -32,6 +43,36 @@ export const BasicModal = () => {
   );
 };
 
-BasicModal.story = {
-  name: "Default"
+export const Link = () => {
+  const [visible, setVisible] = useState(false);
+  const handleOnButtonClick = () => {
+    setVisible(true);
+  };
+  const handleOnClose = () => {
+    setVisible(false);
+  };
+  const handleOnConfirm = () => {
+    setVisible(false);
+  };
+  return (
+    <>
+      <StyledLink onClick={handleOnButtonClick}>
+        Function Name Lorem Ipsum &gt; Agile Group &gt; Breadcrumb
+      </StyledLink>
+      <Modal
+        visible={visible}
+        onConfirm={handleOnConfirm}
+        onClose={handleOnClose}
+        width="50%"
+      >
+        <Spacer my="r">
+          <H2>Agile Group Information</H2>
+          <Divider light />
+          <Box bg="greyLightest" p="r">
+            <P>Lorem ipsum dolor sit amet</P>
+          </Box>
+        </Spacer>
+      </Modal>
+    </>
+  );
 };
